tests: deduplicate sst_importer and resolved_ts test code (#16594)
ref tikv/tikv#16533

* Remove duplicated sst importer test utility functions.
* Remove duplicated resolved ts endpoint initialization code.
