+++ tests/failpoints/cases/test_import_service.rs
@@ -6,8 +6,8 @@ use std::{
 };
 
 use file_system::calc_crc32;
-use futures::{executor::block_on, stream, SinkExt};
-use grpcio::{ChannelBuilder, Environment, Result, WriteFlags};
+use futures::executor::block_on;
+use grpcio::{ChannelBuilder, Environment};
 use kvproto::{import_sstpb::*, tikvpb_grpc::TikvClient};
 use tempfile::{Builder, TempDir};
 use test_raftstore::{must_raw_put, Simulator};
@@ -18,9 +18,10 @@ use tikv_util::{config::ReadableSize, HandyRwLock};
 #[allow(dead_code)]
 #[path = "../../integrations/import/util.rs"]
 mod util;
+
 use self::util::{
-    check_ingested_kvs, new_cluster_and_tikv_import_client, new_cluster_and_tikv_import_client_tde,
-    open_cluster_and_tikv_import_client_v2, send_upload_sst,
+    new_cluster_and_tikv_import_client, new_cluster_and_tikv_import_client_tde,
+    open_cluster_and_tikv_import_client_v2,
 };
 
 // Opening sst writer involves IO operation, it may block threads for a while.
@@ -63,31 +64,8 @@ fn test_download_sst_blocking_sst_writer() {
     fail::remove(sst_writer_open_fp);
 
     // Do an ingest and verify the result is correct.
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx.clone());
-    ingest.set_sst(meta);
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error());
-
+    must_ingest_sst(&import, ctx.clone(), meta);
     check_ingested_kvs(&tikv, &ctx, sst_range);
-}
-
-fn upload_sst(import: &ImportSstClient, meta: &SstMeta, data: &[u8]) -> Result<UploadResponse> {
-    let mut r1 = UploadRequest::default();
-    r1.set_meta(meta.clone());
-    let mut r2 = UploadRequest::default();
-    r2.set_data(data.to_vec());
-    let reqs: Vec<_> = vec![r1, r2]
-        .into_iter()
-        .map(|r| Result::Ok((r, WriteFlags::default())))
-        .collect();
-    let (mut tx, rx) = import.upload().unwrap();
-    let mut stream = stream::iter(reqs);
-    block_on(async move {
-        tx.send_all(&mut stream).await?;
-        tx.close().await?;
-        rx.await
-    })
 }
 
 #[test]
@@ -104,11 +82,7 @@ fn test_ingest_reentrant() {
     let (mut meta, data) = gen_sst_file(sst_path, sst_range);
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
-    upload_sst(&import, &meta, &data).unwrap();
-
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx);
-    ingest.set_sst(meta.clone());
+    send_upload_sst(&import, &meta, &data).unwrap();
 
     // Don't delete ingested sst file or we cannot find sst file in next ingest.
     fail::cfg("dont_delete_ingested_sst", "1*return").unwrap();
@@ -124,9 +98,8 @@ fn test_ingest_reentrant() {
         .get_path(&meta);
 
     let checksum1 = calc_crc32(save_path.clone()).unwrap();
-    // Do ingest and it will ingest successs.
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error());
+    // Do ingest and it will ingest success.
+    must_ingest_sst(&import, ctx.clone(), meta.clone());
 
     let checksum2 = calc_crc32(save_path).unwrap();
     // TODO: Remove this once write_global_seqno is deprecated.
@@ -134,8 +107,7 @@ fn test_ingest_reentrant() {
     // updated with the default setting, which is write_global_seqno=false.
     assert_eq!(checksum1, checksum2);
     // Do ingest again and it can be reentrant
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error());
+    must_ingest_sst(&import, ctx.clone(), meta);
 }
 
 #[test]
@@ -153,7 +125,7 @@ fn test_ingest_key_manager_delete_file_failed() {
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
 
-    upload_sst(&import, &meta, &data).unwrap();
+    send_upload_sst(&import, &meta, &data).unwrap();
 
     let deregister_fp = "key_manager_fails_before_delete_file";
     // the first delete is in check before ingest, the second is in ingest cleanup
@@ -164,12 +136,7 @@ fn test_ingest_key_manager_delete_file_failed() {
     // Do an ingest and verify the result is correct. Though the ingest succeeded,
     // the clone file is still in the key manager
     // TODO: how to check the key manager contains the clone key
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx.clone());
-    ingest.set_sst(meta.clone());
-    let resp = import.ingest(&ingest).unwrap();
-
-    assert!(!resp.has_error());
+    must_ingest_sst(&import, ctx.clone(), meta.clone());
 
     fail::remove(deregister_fp);
 
@@ -193,12 +160,8 @@ fn test_ingest_key_manager_delete_file_failed() {
 
     // Do upload and ingest again, though key manager contains this file, the ingest
     // action should success.
-    upload_sst(&import, &meta, &data).unwrap();
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx);
-    ingest.set_sst(meta);
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error());
+    send_upload_sst(&import, &meta, &data).unwrap();
+    must_ingest_sst(&import, ctx, meta);
 }
 
 #[test]
@@ -215,7 +178,7 @@ fn test_ingest_file_twice_and_conflict() {
     let (mut meta, data) = gen_sst_file(sst_path, sst_range);
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
-    upload_sst(&import, &meta, &data).unwrap();
+    send_upload_sst(&import, &meta, &data).unwrap();
     let mut ingest = IngestRequest::default();
     ingest.set_context(ctx);
     ingest.set_sst(meta);
@@ -267,16 +230,11 @@ fn test_delete_sst_v2_after_epoch_stale() {
     // disable data flushed
     fail::cfg("on_flush_completed", "return()").unwrap();
     send_upload_sst(&import, &meta, &data).unwrap();
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx.clone());
-    ingest.set_sst(meta.clone());
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
     send_upload_sst(&import, &meta, &data).unwrap();
-    ingest.set_sst(meta.clone());
+    must_ingest_sst(&import, ctx.clone(), meta.clone());
 
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error(), "{:?}", resp.get_error());
     let (tx, rx) = channel::<()>();
     let tx = Arc::new(Mutex::new(tx));
     fail::cfg_callback("on_cleanup_import_sst_schedule", move || {
@@ -333,15 +291,10 @@ fn test_delete_sst_after_applied_sst() {
     let (mut meta, data) = gen_sst_file(sst_path, sst_range);
     // No region id and epoch.
     send_upload_sst(&import, &meta, &data).unwrap();
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx.clone());
-    ingest.set_sst(meta.clone());
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
     send_upload_sst(&import, &meta, &data).unwrap();
-    ingest.set_sst(meta.clone());
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error(), "{:?}", resp.get_error());
+    must_ingest_sst(&import, ctx.clone(), meta);
 
     // restart node
     cluster.stop_node(1);
@@ -391,16 +344,10 @@ fn test_split_buckets_after_ingest_sst_v2() {
     let sst_range = (0, 255);
     let (mut meta, data) = gen_sst_file(sst_path, sst_range);
     send_upload_sst(&import, &meta, &data).unwrap();
-    let mut ingest = IngestRequest::default();
-    ingest.set_context(ctx.clone());
-    ingest.set_sst(meta.clone());
     meta.set_region_id(ctx.get_region_id());
     meta.set_region_epoch(ctx.get_region_epoch().clone());
     send_upload_sst(&import, &meta, &data).unwrap();
-    ingest.set_sst(meta.clone());
-
-    let resp = import.ingest(&ingest).unwrap();
-    assert!(!resp.has_error(), "{:?}", resp.get_error());
+    must_ingest_sst(&import, ctx.clone(), meta);
 
     let (tx, rx) = channel::<()>();
     let tx = Arc::new(Mutex::new(tx));
@@ -475,15 +422,10 @@ fn test_flushed_applied_index_after_ingset() {
         let (mut meta, data) = gen_sst_file(sst_path.clone(), sst_range);
         // No region id and epoch.
         send_upload_sst(&import, &meta, &data).unwrap();
-        let mut ingest = IngestRequest::default();
-        ingest.set_context(ctx.clone());
-        ingest.set_sst(meta.clone());
         meta.set_region_id(ctx.get_region_id());
         meta.set_region_epoch(ctx.get_region_epoch().clone());
         send_upload_sst(&import, &meta, &data).unwrap();
-        ingest.set_sst(meta.clone());
-        let resp = import.ingest(&ingest).unwrap();
-        assert!(!resp.has_error(), "{:?}", resp.get_error());
+        must_ingest_sst(&import, ctx.clone(), meta);
     }
 
     // only 1 sst left because there is no more event to trigger a raft ready flush.
@@ -495,15 +437,10 @@ fn test_flushed_applied_index_after_ingset() {
         let (mut meta, data) = gen_sst_file(sst_path.clone(), sst_range);
         // No region id and epoch.
         send_upload_sst(&import, &meta, &data).unwrap();
-        let mut ingest = IngestRequest::default();
-        ingest.set_context(ctx.clone());
-        ingest.set_sst(meta.clone());
         meta.set_region_id(ctx.get_region_id());
         meta.set_region_epoch(ctx.get_region_epoch().clone());
         send_upload_sst(&import, &meta, &data).unwrap();
-        ingest.set_sst(meta.clone());
-        let resp = import.ingest(&ingest).unwrap();
-        assert!(!resp.has_error(), "{:?}", resp.get_error());
+        must_ingest_sst(&import, ctx.clone(), meta);
     }
 
     // ingest more sst files, unflushed index still be 1.
